# backend/services/gateway/api/openapi.yaml
openapi: 3.0.3
info:
  title: Platform API Gateway
  version: 1.0.0
  description: |
    Unified API Gateway for the multi-tenant cloud deployment platform.
    All frontend traffic is routed through this gateway which handles:
    - Authentication (JWT and service account tokens)
    - Authorization (organization and project scope)
    - Rate limiting
    - Request correlation
    - Distributed tracing
servers:
  - url: http://localhost:8080/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /organizations:
    post:
      tags: [Organizations]
      operationId: createOrganization
      summary: Create organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "201":
          description: Organization created
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /organizations/{orgId}:
    parameters:
      - $ref: "#/components/parameters/OrgId"
    get:
      tags: [Organizations]
      operationId: getOrganization
      summary: Get organization
      responses:
        "200":
          description: Organization details
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Organizations]
      operationId: updateOrganization
      summary: Update organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationRequest"
      responses:
        "200":
          description: Organization updated
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Organizations]
      operationId: deleteOrganization
      summary: Delete organization
      responses:
        "204":
          description: Organization deleted
        "403":
          $ref: "#/components/responses/Forbidden"
  /organizations/{orgId}/members:
    parameters:
      - $ref: "#/components/parameters/OrgId"
    get:
      tags: [Members]
      operationId: listMembers
      summary: List organization members
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Member list
  /organizations/{orgId}/members/{userId}:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - $ref: "#/components/parameters/UserId"
    patch:
      tags: [Members]
      operationId: changeRole
      summary: Change member role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeRoleRequest"
      responses:
        "200":
          description: Role changed
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Members]
      operationId: removeMember
      summary: Remove member
      responses:
        "204":
          description: Member removed
        "403":
          $ref: "#/components/responses/Forbidden"
  /organizations/{orgId}/invitations:
    parameters:
      - $ref: "#/components/parameters/OrgId"
    post:
      tags: [Invitations]
      operationId: inviteMember
      summary: Invite member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteMemberRequest"
      responses:
        "201":
          description: Invitation sent
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      tags: [Invitations]
      operationId: listInvitations
      summary: List pending invitations
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Invitation list
  /organizations/{orgId}/invitations/{id}:
    parameters:
      - $ref: "#/components/parameters/OrgId"
      - name: id
        in: path
        required: true
        schema:
          type: string
    delete:
      tags: [Invitations]
      operationId: revokeInvitation
      summary: Revoke invitation
      responses:
        "204":
          description: Invitation revoked
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token issued by the auth service
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Organization ID
...
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: User ID
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Maximum number of items to return
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Pagination cursor from previous response
  schemas:
...
    CreateOrganizationRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 64
        slug:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]*[a-z0-9]$"
      required: [name, slug]
    UpdateOrganizationRequest:
      type: object
      properties:
        name:
          type: string
        plan:
          type: string
    InviteMemberRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, developer, viewer]
      required: [email, role]
...
    ChangeRoleRequest:
      type: object
      properties:
        role:
          type: string
          enum: [owner, admin, developer, viewer]
      required: [role]
...
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        requestId:
          type: string
      required: [error]
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...